# feat: add `Validation.SuccessWithFailures`