--- a/main/src/ca/uwaterloo/flix/language/phase/Weeder.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/Weeder.scala
@@ -1918,6 +1918,7 @@ object Weeder {
               // Doing a manual flatMap to keep the function tail-recursive
               translateHexCode(code, mkSL(sp1, sp2)) match {
                 case Validation.Success(char) => visit(rest2, char :: acc)
+                case Validation.SuccessWithFailures(_, _) => ???
                 case Validation.Failure(errors) => Validation.Failure(errors)
               }
             // Case 3.3.2: `\\u` followed by less than 4 literals