--- a/main/src/ca/uwaterloo/flix/tools/SimpleRunner.scala
+++ b/main/src/ca/uwaterloo/flix/tools/SimpleRunner.scala
@@ -116,6 +116,8 @@ object SimpleRunner {
 
         ().toOk
 
+      case Validation.SuccessWithFailures(_, _) => ???
+
       case Validation.Failure(errors) =>
         flix.mkMessages(errors.sortBy(_.source.name))
           .foreach(println)