--- a/main/src/ca/uwaterloo/flix/util/Validation.scala
+++ b/main/src/ca/uwaterloo/flix/util/Validation.scala
@@ -205,6 +205,7 @@ object Validation {
                     (f: T1 => U): Validation[U, E] =
     t1 match {
       case Success(v1) => Success(f(v1))
+      case SuccessWithFailures(v1, err1) => SuccessWithFailures(f(v1), err1)
       case _ => Failure(t1.errors)
     }