Supersedes https://github.com/flix/flix/pull/5144

Related to https://github.com/flix/flix/issues/5026