@jaschdoc 2022-12-24
I've added `ap` and `curry` functions to the validation functions to keep the implementation of `mapN` simple. It may be a bit slower, however, the combinatorial explosion of each `mapN` implementation would be both very time consuming to read and write.