@jaschdoc 2022-12-24
# Benchmarks

**It looks like there is no significant performance decrease.**

## This branch

The first result is slightly off. I had other programs running during this benchmark so maybe it could have affected the result.
```
====================== Flix Compiler Throughput ======================

Throughput (best): 12,188 lines/sec (with 8 threads.)

  min:  3,973, max: 12,188, avg:  9,558, median:  9,884

  The highest throughput was in iteration: 8 (out of 10).
  The ratio between the best and worst iteration was: 3.1x.

Finished 10 iterations on 62,768 lines of code in 72 seconds.

                                                                                
====================== Flix Compiler Throughput ======================

Throughput (best): 13,214 lines/sec (with 8 threads.)

  min:  4,352, max: 13,214, avg: 10,713, median: 11,359

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 3.0x.

Finished 10 iterations on 62,768 lines of code in 64 seconds.


====================== Flix Compiler Throughput ======================

Throughput (best): 12,787 lines/sec (with 8 threads.)

  min:  4,495, max: 12,787, avg: 10,259, median: 11,112

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 2.8x.

Finished 10 iterations on 62,768 lines of code in 66 seconds.
```

## Main

```
====================== Flix Compiler Throughput ======================

Throughput (best): 12,871 lines/sec (with 8 threads.)

  min:  4,731, max: 12,871, avg: 10,400, median: 11,610

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 2.7x.

Finished 10 iterations on 62,768 lines of code in 65 seconds.

====================== Flix Compiler Throughput ======================

Throughput (best): 12,222 lines/sec (with 8 threads.)

  min:  4,518, max: 12,222, avg: 10,220, median: 10,767

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 2.7x.

Finished 10 iterations on 62,768 lines of code in 66 seconds.

====================== Flix Compiler Throughput ======================

Throughput (best): 12,578 lines/sec (with 8 threads.)

  min:  4,432, max: 12,578, avg: 10,478, median: 11,560

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 2.8x.

Finished 10 iterations on 62,768 lines of code in 65 seconds.
```