@magnus-madsen 2022-12-24
Ok. We could also just specialise the ones for 1 to 3 cases.