@jaschdoc 2022-12-27
> RFR?

Almost, just need tests