@jaschdoc 2022-12-27
> OK. Please also build a JAR and verify that it works correctly with VSCode when writing a program with/without errors :)

Roger!