@magnus-madsen 2022-12-27
@mlutze Can I ask you to do a careful look through `Validation.scala` to see that we are not missing any errors?