@jaschdoc 2022-12-27
> Will you "unspecialize" the special cases and run another benchmark throughput?

Sure :)