@magnus-madsen 2022-12-27
Two things left:
- The combineFailureStreams thing
- Benchmarks