@jaschdoc 2022-12-27
# Benchmarks

## This branch f2baf2a511cd9e207cf3c2915727d7d42336b170

```
====================== Flix Compiler Throughput ======================

Throughput (best): 11,150 lines/sec (with 8 threads.)

  min:  4,273, max: 11,150, avg:  8,872, median:  9,096

  The highest throughput was in iteration: 6 (out of 10).
  The ratio between the best and worst iteration was: 2.6x.

Finished 10 iterations on 62,779 lines of code in 76 seconds.

                                                                                
====================== Flix Compiler Throughput ======================

Throughput (best): 11,026 lines/sec (with 8 threads.)

  min:  3,648, max: 11,026, avg:  9,096, median: 10,080

  The highest throughput was in iteration: 5 (out of 10).
  The ratio between the best and worst iteration was: 3.0x.

Finished 10 iterations on 62,779 lines of code in 76 seconds.


====================== Flix Compiler Throughput ======================

Throughput (best): 12,435 lines/sec (with 8 threads.)

  min:  3,696, max: 12,435, avg:  9,969, median: 10,439

  The highest throughput was in iteration: 8 (out of 10).
  The ratio between the best and worst iteration was: 3.4x.

Finished 10 iterations on 62,779 lines of code in 71 seconds.
```

## Main c9dcf3b7b3fe3c7ebd8b6e5ecce9481077a3fe85

```                                                                                
====================== Flix Compiler Throughput ======================

Throughput (best): 13,137 lines/sec (with 8 threads.)

  min:  4,786, max: 13,137, avg: 10,707, median: 11,672

  The highest throughput was in iteration: 7 (out of 10).
  The ratio between the best and worst iteration was: 2.7x.

Finished 10 iterations on 62,779 lines of code in 63 seconds.

                                                                                
====================== Flix Compiler Throughput ======================

Throughput (best): 12,856 lines/sec (with 8 threads.)

  min:  4,755, max: 12,856, avg: 10,561, median: 11,597

  The highest throughput was in iteration: 8 (out of 10).
  The ratio between the best and worst iteration was: 2.7x.

Finished 10 iterations on 62,779 lines of code in 64 seconds.


====================== Flix Compiler Throughput ======================

Throughput (best): 11,464 lines/sec (with 8 threads.)

  min:  3,902, max: 11,464, avg:  9,268, median: 10,394

  The highest throughput was in iteration: 9 (out of 10).
  The ratio between the best and worst iteration was: 2.9x.

Finished 10 iterations on 62,779 lines of code in 74 seconds.
```