@magnus-madsen 2022-12-27
Ill merge; next up, I think there is some dead code and we need to look at performance.