@jaschdoc 2022-12-27
Hadn't benchmarked the latest commit. I think maybe we gained some performance in Matt's version of `fastTraverse`.